fix(web): move paths to its own file
To avoid circular dependencies.
