+++ web/src/components/network/ConnectionsTable.tsx
@@ -25,8 +25,8 @@ import { useNavigate, generatePath } from "react-router-dom";
 import { Table, Thead, Tr, Th, Tbody, Td } from "@patternfly/react-table";
 import { RowActions } from "~/components/core";
 import { Icon } from "~/components/layout";
-import { PATHS } from "~/routes/network";
 import { Connection, Device } from "~/types/network";
+import { NETWORK as PATHS } from "~/routes/paths";
 import { formatIp } from "~/utils/network";
 import { sprintf } from "sprintf-js";
 import { _ } from "~/i18n";
