+++ web/src/components/software/SoftwarePage.tsx
@@ -36,7 +36,7 @@ import { useIssues } from "~/queries/issues";
 import { usePatterns, useProposal, useProposalChanges } from "~/queries/software";
 import { Pattern, SelectedBy } from "~/types/software";
 import { _ } from "~/i18n";
-import { PATHS } from "~/routes/software";
+import { SOFTWARE as PATHS } from "~/routes/paths";
 
 /**
  * List of selected patterns.
