+++ web/src/components/storage/ProposalActionsSummary.tsx
@@ -24,13 +24,13 @@ import React from "react";
 import { Button, Skeleton, Stack, List, ListItem } from "@patternfly/react-core";
 import { Link, Page } from "~/components/core";
 import DevicesManager from "~/components/storage/DevicesManager";
-import { _, n_ } from "~/i18n";
-import { sprintf } from "sprintf-js";
 import textStyles from "@patternfly/react-styles/css/utilities/Text/text";
-import { PATHS } from "~/routes/storage";
 import { Action, SpaceAction, StorageDevice } from "~/types/storage";
 import { SpacePolicy } from "./utils";
 import { ValidationError } from "~/types/issues";
+import { STORAGE as PATHS } from "~/routes/paths";
+import { sprintf } from "sprintf-js";
+import { _, n_ } from "~/i18n";
 
 /**
  * Renders information about delete actions
